Switch analysis.FindDistance return type to int from int64

+++ internal/analysis/metrics.go
@@ -5,10 +5,10 @@ import (
 	"math"
 )
 
-func FindDistance(from, to *stapi.Location) int64 {
+func FindDistance(from, to *stapi.Location) int {
 	x := float64(to.XCoordinate - from.XCoordinate)
 	y := float64(to.YCoordinate - from.YCoordinate)
-	return int64(math.Round(math.Sqrt(x*x + y*y)))
+	return int(math.Round(math.Sqrt(x*x + y*y)))
 }
 
 const fuelContingency = 2
